requirement.txt
affine==2.4.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
fiona==1.9.6
fonttools==4.50.0
GDAL==3.8.4
geopandas==0.14.3
geoserver-rest==2.6.0
greenlet==3.0.3
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
postgres-helper==0.0.5
psycopg2==2.9.9
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.3
PyOpenGL==3.1.7
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
rasterio==1.3.9
requests==2.31.0
Rtree==1.2.0
seaborn==0.13.2
setuptools==69.2.0
shapely==2.0.3
six==1.16.0
snuggs==1.4.7
SQLAlchemy==2.0.29
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
xmltodict==0.13.0
